+++ src/lib/ui/Forum_Items.svelte
@@ -1,7 +1,9 @@
 <script lang="ts">
 	import type {Post} from '$lib/posts/post.js';
 	import type {Member} from '$lib/members/member.js';
-	import Board_List_Item from '$lib/ui/Board_List_Item.svelte';
+	import Forum_Item from '$lib/ui/Forum_Item.svelte';
+
+	// TODO this should possibly be a generic component instead of this named one
 
 	export let posts: Post[];
 	export let members_by_id: Map<number, Member>;
@@ -17,6 +19,6 @@
 <!-- TODO possibly remove the `ul` wrapper and change the `li`s to `div`s -->
 <ul>
 	{#each posts as post (post.post_id)}
-		<Board_List_Item {post} member={to_member(post.actor_id)} />
+		<Forum_Item {post} member={to_member(post.actor_id)} />
 	{/each}
 </ul>
