+++ src/lib/ui/Notes_Item.svelte
